std.random2.generator module first entry: MersenneTwisterEngine.
This is a class-based reimplementation which introduces extra
template parameters found in the latest version of boost::random.
This enables us to implement the 64-bit Mt19937 as well as the
more common 32-bit one.

Seeding is currently limited to single seed values, with range-
based seeds to be added later.