TypeTuple containing all uniform RNGs defined in the module.
The goal here is to offer a useful tool for testing, both
internally in the module and externally for any users writing
code designed to work with RNGs.  So far as internal testing
goes, this allows us to separate out general tests which all
uniform RNGs should pass from tests specific to a particular
generator.