Add UniformDistribution for [T.min, T.max] integral variates
This adds UniformDistribution support for its uniform!(T, UniformRNG)
counterpart.

Note that one potential bug in the current implementation is the fact
that when templated on a character type, it is possible that invalid
UTF8 code points will be generated, which may lead to segfaults or
errors when e.g. writing to console or comparing character values.