Ensure bounded UniformDistribution supports all int/char types
The cast for the lower bound in uniform() has been corrected: the
earlier check that a < ResultType.max should ensure that there are no
problems with this.

Unittests have been added to ensure that all types are compatible with
all distribution boundaries.

The uniformDistribution helper function arguments have been converted
from ref UniformRNG to UniformRNG to allow rvalue RNG instances to be
passed.  As this package presupposes all RNGs are reference types,
retaining support for ref UniformRNG parameters does not seem necessary.