Improved unittests and experimental-status alerts for random devices
The random device module has been removed from the main package due to
its experimental status, however it can still be imported as a module in
its own right.

The unittests have been improved and made silent.  Note that they may
take an extended period of time to run (over a minute) due to the
blocking nature of /dev/random.