Convert hap to a dub package
The dub package setup divides hap.random into a main package consisting
of all the non-experimental parts of the library, and the subpackage
hap.random:experimental which (as the name suggests:-) contains the
experimental parts, currently limited to hap.random.device.

The Makefile is preserved, but has been tweaked to deal with the new
directory structure.  Conceptually one should see the Makefile as
serving the needs of the developer hacking directly on hap.random,
providing unittest and benchmark support, while the dub package supports
3rd-party use.

A version number has been omitted for now, and will be added after final
tweaks are made to the library.